requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
psycopg2-binary==2.9.9
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
python-multipart==0.0.20
requests==2.32.4
sqlalchemy==2.0.23
pyjwt==2.8.0
geopandas==0.14.1
fiona==1.9.5
shapely==2.0.2
pyproj==3.6.1
